refactor: file-input을 input-image로 파일명 변경
