+++ packages/ui/src/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./ui/button";
-export * from "./ui/file-input";
 export * from "./ui/input-field";
+export * from "./ui/input-image";
 export * from "./ui/input-text-area";
