+++ packages/ui/src/components/ui/input-image.tsx
